requirements.txt
Flask==3.0.3
Flask-Cors>=4.0.1
python-dotenv>=1.0.1
supabase==2.6.0
httpx>=0.27.0,<0.30.0
requests>=2.31.0
openai>=1.109.1
pydantic>=2.11.9
uvicorn>=0.30.6
crewai==1.3.0
crewai-tools==1.3.0
tavily-python
pytest==8.4.2
beautifulsoup4
lxml==5.3.0
stripe>=10.0.0